update feature books item and listview
